Struct wasm_streams::readable::PipeOptions

pub struct PipeOptions { /* private fields */ }

Options for pipe_to_with_options.

Implementations

impl PipeOptions

pub fn new() -> Self

Creates a blank new set of pipe options.

Equivalent to PipeOptions::default.

pub fn from_raw(raw: PipeOptions) -> Self

Creates a set of pipe options from a raw PipeOptions object.

pub fn into_raw(self) -> PipeOptions

Convert this to a raw PipeOptions object.

pub fn prevent_close(&mut self, prevent_close: bool) -> &mut Self

Sets whether the destination writable stream should be closed when the source readable stream closes.

pub fn prevent_cancel(&mut self, prevent_cancel: bool) -> &mut Self

Sets whether the source readable stream should be canceled when the destination writable stream errors.

pub fn prevent_abort(&mut self, prevent_abort: bool) -> &mut Self

Sets whether the destination writable stream should be aborted when the source readable stream errors.

pub fn signal(&mut self, signal: AbortSignal) -> &mut Self

Sets an abort signal to abort the ongoing pipe operation. When the signal is aborted, the source readable stream will be canceled and the destination writable stream will be aborted unless the respective options prevent_cancel or prevent_abort are set.